test: create filter robot
